backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.7.1
httpx==0.27.0
python-multipart==0.0.9
beautifulsoup4>=4.12.0
lxml>=4.9.0
Pillow>=10.0.0
